Moves default implention of snapshot directory and id to concerns
