+++ src/Concerns/SnapshotDirectoryAware.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Spatie\Snapshots\Concerns;
+
+use ReflectionClass;
+
+trait SnapshotDirectoryAware
+{
+    /*
+     * Determines the directory where snapshots are stored. By default a
+     * `__snapshots__` directory is created at the same level as the test
+     * class.
+     */
+    protected function getSnapshotDirectory(): string
+    {
+        return dirname((new ReflectionClass($this))->getFileName()).
+            DIRECTORY_SEPARATOR.
+            '__snapshots__';
+    }
+}
